Fix: insights data not send
