+++ wpuf.php
@@ -150,6 +150,9 @@ final class WP_User_Frontend {
         // set schedule event
         add_action( 'wpuf_remove_expired_post_hook', array( $this, 'action_to_remove_exipred_post' ) );
         add_action( 'wp_ajax_wpuf_weforms_install', array( $this, 'install_weforms' ) );
+
+        // Insight class instentiate
+        $this->container['insights'] = new WPUF_WeDevs_Insights( __FILE__ );
     }
 
     /**
@@ -328,7 +331,6 @@ final class WP_User_Frontend {
         $this->container['subscription']            = WPUF_Subscription::init();
         $this->container['frontend_post']           = WPUF_Frontend_Form_Post::init();
         $this->container['account']                 = new WPUF_Frontend_Account();
-        $this->container['insights']                = new WPUF_WeDevs_Insights( 'wp-user-frontend', 'WP User Frontend', __FILE__ );
         $this->container['billing_address']         = new WPUF_Ajax_Address_Form();
         $this->container['forms']                   = new WPUF_Form_Manager();
         $this->container['preview']                 = new WPUF_Form_Preview();
